Adding hooks for serialization

+++ blog/middleware/redis.py
@@ -11,16 +11,16 @@ class CacheProvider(object):
 
     def process_request(self, req, resp):
         """Provide redis cache with every request."""
-        req.context.set('cached', client.get(req.uri, None))
+        req.context.set('cached', client.get(req.uri))
 
     def process_response(self, req, resp, resource, req_succeeded):
         """Sets or deletes cache for provided resources."""
-        if req_succeeded:
+        if req_succeeded and resource.use_cache:
             if req.method == HttpMethods.GET:
                 client.set(req.uri, resp.body)
             else:
                 client.delete(req.uri)
                 for resc in resource.cached_resources:
-                    route = resource.route.replace('{', '${')  # interpolate for safe formatting
+                    route = resc.route.replace('{', '${')  # interpolate for safe formatting
                     # assumes that binded resources ay have routes with similar params
                     client.delete(Template(route).safe_substitute(**req.params))
